Changes FASTA download buttons
The text of the FASTA download buttons has been reworded. In addition, the
species name has been removed from the same species FASTA download buttons,
as the species is the same in both buttons.

Finally, a new FASTA download button for downloading a FASTA with all the
genes in both interactomes of a same species result has been added.