Isolates results panel from form panel
The results panel has been redesigned to use only data coming from the
query result. This requires more calls to the server, but it makes it
more reusable.

This change has allowed an improvement in the application when accessing
to the results using the permalinks. Now, the application, in these
cases, behaves as with normal queries.