Fixes BLAST params order and limits
The "Number of descriptors" and the "Minimum length of alignment block"
sliders were crossed. This commit places the sliders next to their
corresponding inputs.

In addition, the input fields didn't have the same limits as the
sliders. In this commit, the same limits have been set to each pair
of input and sliders.