Fixes Same Species form parsing of results
Each interaction supports multiple interactomes instead of just two.
Interactomes are now selected with one control.
In the graph will be represented with 3 link styles if the interaction
includes one interactome, all interactomes or the rest of combinations.
Modifies the table and CSV to include one column for eath interactome
selected.
Download of interactomes in FASTA format is displayed in a menu.