Improves usability of the query forms
Usability of the query forms have been improved by disabling the fields
until they can be used. For example, interactomes field is disabled
until the corresponding species has been selected.

In addition, the "go to results" message is now shown in a confirmation
bottom sheet and the form is cleared after a request is successfully
done.