Fixes error with results without steps
Although a same or different species result should always have, at
least, one step, an error during their execution may cause the backend
to serve a result with an empty step list.

This was causing the frontend to fail, as a reduce operation was applied
to the steps array in order to find the last step and this operation
fails when the array is empty and no initial value is provided.

This commit fixes this error by providing a initial value to the reduce
operation.