Improves result table views
The result table views have been improved in several ways.

In both views, the degree presentation in the table is now more
simple ("Degree" work has been removed), but more informative, as for
collapsed interactomes each degree includes the number of
interactomes with that degree and a tooltip to view the interactomes.

In addition, the code has been simplified and some unnecessary parts
have been removed.

In the same species view, a new option to filter the interactomes
show has been added and the CSV preparation now supports four
different modes that can be choosen when the CSV is prepared.