Fixes some errors in interactome/predictome lists
The lists were not updated when a query returned by the backend was
empty due to the forkJoin behaviour in the corresponding services.
The services were fixed and now it works properly.

Some columns missed the filtering field, so a filter field was added
to them.

Finally, some minor typos and code formating errors wera also fixed.