Adds upgrade notice to results
When the production application is updated the old results will be
removed, as they are not compatible with the new version.

To prevent users from losing their resutls, the old version will be
available for some time in an alternative address. Therefore, an
upgrading notice to inform users about this was added to the results
view.